refactor: move configuration assessment constants to a separate file for better organization
