+++ plugins/wazuh-analysis/public/groups/endpoint-security/applications.ts
@@ -6,13 +6,15 @@ import {
   AppUpdater,
 } from '../../../../../src/core/public';
 import {
-  CONFIGURATION_ASSESSMENT_ID,
-  CONFIGURATION_ASSESSMENT_TITLE,
   FIM_ID,
   FIM_TITLE,
   MALWARE_DETECTION_ID,
   MALWARE_DETECTION_TITLE,
 } from './constants';
+import {
+  CONFIGURATION_ASSESSMENT_ID,
+  CONFIGURATION_ASSESSMENT_TITLE,
+} from './apps/configuration-assesment/constants';
 
 export function getEndpointSecurityApps(updater$?: Subject<AppUpdater>): App[] {
   return [
